requirements-base.txt
django>1.6
Pillow>=2
django-extensions
django-classy-tags>=0.5
South>=0.8.4
html5lib>=0.90
django-mptt==0.6
django-sekizai>=0.7
argparse
dj-database-url
djangocms-admin-style
https://github.com/KristianOellegaard/django-hvad/archive/master.zip
djangocms-link
https://github.com/stefanfoulis/django-filer/archive/develop.zip
django-reversion>=1.8
https://github.com/stefanfoulis/cmsplugin-filer/archive/develop.zip
djangocms-blog>=0.2b5
djangocms-style
django-admin-sortable2
